+++ WebKalkulacka/WebProject/Models/Calculation.cs
@@ -8,10 +8,10 @@ namespace WebProject.Models
         [Key]
         public int Id { get; set; }
 
-        public double Number1 { get; set; }
-        public double Number2 { get; set; }
+        public decimal Number1 { get; set; }
+        public decimal Number2 { get; set; }
         public Operation Operation { get; set; }
-        public double Result { get; set; }
+        public decimal Result { get; set; }
 
     }
 }
